= optimize obj structure
